everything is not in a good state

+++ ApiDemo/DataBase/Classes/FileServer.cs
@@ -3,10 +3,10 @@ using ApiDemo.DataBase.Interfaces;
 namespace ApiDemo.DataBase.Classes;
 
 public class FileServer : IFileServer {
-    public void UploadFile(string filePath, string fileName, Stream stream) {
-        string completePath = Path.Combine(filePath, fileName);
+    public async Task UploadFile(string filePath, string fileName, Stream stream) {
+        var completePath = Path.Combine(filePath, fileName);
         using var fileStream = File.Create(completePath);
-        stream.CopyTo(fileStream);
+        await stream.CopyToAsync(fileStream);
         fileStream.Close();
     }
 
@@ -15,7 +15,7 @@ public class FileServer : IFileServer {
     }
 
     public void DeleteFile(string filePath, string fileName) {
-        string completePath = Path.Combine(filePath, fileName);
+        var completePath = Path.Combine(filePath, fileName);
         File.Delete(completePath);
     }
 
@@ -24,13 +24,11 @@ public class FileServer : IFileServer {
     }
 
     public Stream GetFile(string filePath, string fileName) {
-        string completePath = Path.Combine(filePath, fileName);
+        var completePath = Path.Combine(filePath, fileName);
         return GetFile(completePath);
     }
 
     public void DeleteFolder(string path) {
-        if (Directory.Exists(path)) {
-        Directory.Delete(path, true);
-        }
+        if (Directory.Exists(path)) Directory.Delete(path, true);
     }
 }
